refactor: deploy and infra dirs
