+++ deploy/dev/nginx.conf
@@ -64,7 +64,7 @@ http {
     server_name app.open-mfa.local;
     listen 443 ssl;
     location / {
-      proxy_pass http://open-mfa:8080;
+      proxy_pass http://172.19.0.1:8080;
     }
   }
 }
