+++ deploy/docker-compose.yaml
@@ -5,13 +5,15 @@ networks:
       config:
         - subnet: 172.19.0.0/24
 
+volumes:
+  shared-certs:
+
 services:
   open-mfa:
     image: rodzera/open-mfa:latest
     container_name: open-mfa
-    restart: on-failure:3
-    depends_on:
-      - redis
+    restart: unless-stopped
+    depends_on: [redis]
     environment:
       _SECRET_KEY: ${_SECRET_KEY}
       _B64_AES_KEY: ${_B64_AES_KEY}
@@ -21,19 +23,15 @@ services:
       _REDIS_PASS: ${_REDIS_PASS}
       HOST: 0.0.0.0
       PORT: 8080
-      WORKERS: 2
-      THREADS: 4
-      WORKER_CONNECTIONS: 1000
       TZ: Etc/Greenwich
-    networks:
-      open-mfa:
-        aliases:
-          - open-mfa
-        ipv4_address: 172.19.0.2
     ports:
       - "8080:8080"
     volumes:
       - /etc/localtime:/etc/localtime:ro
+    networks:
+      open-mfa:
+        aliases: [open-mfa]
+        ipv4_address: 172.19.0.2
 
   redis:
     image: redis:7.4.1-alpine
@@ -44,10 +42,7 @@ services:
     command: sh -c 'redis-server --requirepass "$_REDIS_PASS"'
     ports:
       - "6379:6379"
-    volumes:
-      - ./redis_data:/data
     networks:
       open-mfa:
-        aliases:
-          - redis
+        aliases: [redis]
         ipv4_address: 172.19.0.3
