--- infra/Makefile
@@ -1,11 +1,0 @@
-dev-up:
-	docker compose -f docker-compose.dev.yaml up -d
-
-dev-down:
-	docker compose -f docker-compose.dev.yaml down
-
-prd-up:
-	docker compose -f docker-compose.yaml up -d
-
-prd-down:
-	docker compose -f docker-compose.yaml down -d
