--- src/Makefile
@@ -1,2 +1,0 @@
-all:
-	docker build --tag=rodzera/open-mfa:dev --tag=rodzera/open-mfa:latest . > docker-build.log
