+++ src/docker/Makefile
@@ -1,0 +1,6 @@
+all:
+	docker build \
+		--file Dockerfile \
+		--tag rodzera/open-mfa:dev \
+		--tag rodzera/open-mfa:latest \
+		..
